requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-filter==23.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2